Moved test classes to /resources/feature folder
