+++ Backend/src/test/java/resources/feature/LeaveTournamentTeamConfirmationEmailDefinitions.java
@@ -1,3 +1,4 @@
+package resources.feature;
 
 import dat.entities.PlayerAccount;
 import dat.entities.Team;
