+++ Backend/src/test/java/resources/feature/LoginStepdefs.java
@@ -1,4 +1,4 @@
-package resources;
+package resources.feature;
 
 import static org.junit.jupiter.api.Assertions.*;
