+++ Backend/src/test/java/resources/feature/UserStrikeStepDefinitions.java
@@ -1,3 +1,5 @@
+package resources.feature;
+
 import dat.config.HibernateConfig;
 import dat.daos.UserDAO;
 import dat.entities.User;
